- Moved define and mount functions to the "public namespace". - Some "final" adjustments to the build tool chain.
