+++ src/component.ts
@@ -1,36 +1,11 @@
+/** @internal */
 import {
-    ComponentFactory,
-    ComponentSetup,
     ComponentVNode,
     UpdateState,
     VNode
 } from './contract'
 import { equal } from './helpers'
 import { render } from './renderer'
-
-/** 
- * Defines a component from a ComponentSpec returning a factory that creates 
- * ComponentVNode/JSXElement's for the component.
- */
-export function define<TState, TProps>(state: TState, spec: ComponentSetup<TState, TProps>) {
-    return function (props: TProps, children: VNode[] = []) {
-        return {
-            _: 3,
-            children,
-            props,
-            state,
-            spec
-        } as any as ComponentVNode<TState, TProps>
-    }
-}
-
-/** 
- * Mounts the component onto the supplied element by calling the supplied 
- * component factory. 
- */
-export function mount(componentFactory: ComponentFactory<any, any>, element: Element) {
-    initComponent(componentFactory({}, []), element)
-}
 
 /** 
  * Initializes a component from a ComponentVNode.
