+++ src/helpers.ts
@@ -1,3 +1,4 @@
+/** @internal */
 export type Type = 'array' | 'object' | 'string' | 'date' | 'regexp' | 'function' | 'boolean' | 'number' | 'null' | 'undefined'
 
 /**
